Implemented async api on service side

+++ SmartDom/SmartDom.Service.Interface/Models/Device.cs
@@ -11,7 +11,7 @@ namespace SmartDom.Service.Interface.Models
 {
     public class Device
     {
-        public ushort Id { get; set; }
+        public byte Id { get; set; }
         public DeviceState State { get; set; }
         public DeviceType Type { get; set; }
         public DeviceSubtype Subtype { get; set; }
